+++ src/game/tests.rs
@@ -604,6 +604,7 @@ mod old_school_upkeep;
 mod one_sided_block_triggers;
 mod open_activation;
 mod osai_vultures;
+mod paralyze;
 mod part_water;
 mod poison;
 mod populate;